Add option to disable create button
At the moment the only way to disable create button is to add a second type (be it fake or otherwise) to the types array. This pull request would give an option of 'createButton' which would default to 'true' but could be set to 'false' if you don't want the intent.

A more explicit way of stopping a button being produced.